+++ tests/testthat/test-write_output.R
@@ -139,9 +139,7 @@ test_that("write_output_dir_structure generates dirs", {
 })
 
 test_that("process_quantiles works as expected", {
-  # Load the sample fit object
-  fit <- readRDS(test_path("data", "sample_fit.rds"))
-
+  # Fit object read in from setup.R
   # Run the function on the fit object
   result <- process_quantiles(
     fit,
@@ -178,7 +176,7 @@ test_that("process_quantiles works as expected", {
   )
 
   # Test 3: Check if the result contains the correct number of rows
-  expected_num_rows <- 55
+  expected_num_rows <- 111
   expect_equal(
     nrow(result),
     expected_num_rows,
@@ -225,9 +223,7 @@ test_that("process_quantiles works as expected", {
 })
 
 test_that("process_samples works as expected", {
-  # Load the sample fit object
-  fit <- readRDS(test_path("data", "sample_fit.rds"))
-
+  # Fit object read in from setup.R
   # Run the function on the fit object
   result <- process_samples(fit, "test_geo", "test_model", "test_disease")
 
@@ -256,7 +252,7 @@ test_that("process_samples works as expected", {
   )
 
   # Test 3: Check if the result contains the correct number of rows
-  expected_num_rows <- 2505 # Replace with actual expected value
+  expected_num_rows <- 1330 # Replace with actual expected value
   expect_equal(
     nrow(result),
     expected_num_rows,